auth requests changed
