+++ graubunden/src/main/java/nu/obama/graubunden/config/SecurityConfig.java
@@ -97,7 +97,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                         "/**/*.css",
                         "/**/*.js")
                 .permitAll()
-                .antMatchers("/api/auth/**")
+                .antMatchers( "/token/**")
                 .permitAll()
                 .antMatchers("/api/user/checkUsernameAvailability", "/api/user/checkEmailAvailability")
                 .permitAll()
@@ -105,6 +105,8 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                         "/api/users/**",
                         "/posts/**",
                         "/groups/**")
+                .permitAll()
+                .antMatchers("/users/**")
                 .permitAll()
                 .anyRequest()
                 .authenticated();
